Refactor. Move fun animationRatingDonut in RatingDonutView.kt
